added UserPayload, UserMapper

+++ src/main/java/com/services/userService/Service/UserServiceImpl.java
@@ -4,8 +4,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.services.userService.interfaces.UserService;
+import com.services.userService.mapper.UserMapper;
 import com.services.userService.models.User;
+import com.services.userService.models.UserPayload;
 import com.services.userService.repository.UserRepository;
+
+import jakarta.persistence.EntityExistsException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -13,28 +18,47 @@ public class UserServiceImpl implements UserService {
     @Autowired
     UserRepository userRepository;
 
-    public User createUser(String email, String password) {
-        User user = new User();
-        user.setEmail(email);
-        user.setPassword(password);
+    @Autowired
+    UserMapper userMapper;
+
+    public UserPayload createUser(UserPayload userPayload) {
+        User user = userMapper.toUser(userPayload);
+        User savedUser = createUser(user);
+        return userMapper.toPayload(savedUser);
+    }
+
+    private User createUser(User user) {
+        User existingUser = userRepository.findByEmail(user.getEmail());
+        if(existingUser!=null){
+            throw new EntityExistsException("User Already Exists");
+        }
         return userRepository.save(user);
     }
 
-    public User getUser(String email) {
-        return userRepository.findByEmail(email);
+    public User getUser(UserPayload userPayload) {
+        User user = userMapper.toUser(userPayload);
+        User existedUser = userRepository.findByEmail(user.getEmail());
+        if(existedUser==null){
+            throw new EntityNotFoundException("User Not Found");
+        }
+        return existedUser;
     }
 
     public Iterable<User> getAllUsers() {
         return userRepository.findAll();
     }
-
-    public User updateUser(String email) {
+    // To be implemented
+    public User updateUser(UserPayload userPayload) {
         return null;
     }
 
-    public void deleteUser(String email) {
-        User user = userRepository.findByEmail(email);
-        userRepository.delete(user);
+    public void deleteUser(UserPayload userPayload) {
+        User user = userMapper.toUser(userPayload);
+        user = userRepository.findByEmail(user.getEmail());
+        if(user==null){
+            throw new EntityNotFoundException("User Not Found");
+        }
+        userRepository.delete(user);  
     }
 
     public void deleteAllUsers() {
